requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.0
Pillow==10.0.0
requests==2.31.0
